backend_flask/requirements.txt
blinker==1.8.2
boto3==1.34.133
botocore==1.34.133
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
ecdsa==0.19.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
load-dotenv==0.1.0
MarkupSafe==2.1.5
pyasn1==0.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
requests==2.32.3
rsa==4.9
s3transfer==0.10.2
six==1.16.0
urllib3==2.2.2
Werkzeug==3.0.3
PyJWT~=2.8.0